rewrite local address detection make some functions static that arr not needed globally rework how fd_set is construction for the main loop select()

+++ src/racoon/isakmp_inf.c
@@ -169,7 +169,7 @@ isakmp_info_recv(iph1, msg0)
 		plog(LLV_ERROR, LOCATION, NULL, 
 			"ignore information because the "
 			"message is way too short - %zu byte(s).\n",
-			(int)msg->l);
+			msg->l);
 		goto end;
 	}
 
@@ -198,7 +198,7 @@ isakmp_info_recv(iph1, msg0)
 			plog(LLV_ERROR, LOCATION, NULL, 
 				"ignore information because the "
 				"message is too short - %zu byte(s).\n",
-				(int)msg->l);
+				msg->l);
 			goto end;
 		}
